Ensure add/remove from collection verifies user visibility

+++ app/Actions/Collection/RemoveDocument.php
@@ -7,6 +7,8 @@ use App\Models\Document;
 use App\Models\Team;
 use App\Models\User;
 use App\Models\Visibility;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Auth\AuthenticationException;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rules\Enum;
@@ -19,11 +21,17 @@ class RemoveDocument
      *
      * @param  array  $input
      */
-    public function __invoke(Document $document, Collection $collection): void
+    public function __invoke(Document $document, Collection $collection, ?User $user = null): void
     {
-        // Gate::forUser($user)->authorize('update', $team);
+        $user = $user ?? auth()->user();
 
-        // TODO: how I can verify that a user is entitled to add a document to a collection
+        if(is_null($user)){
+            throw new AuthenticationException(__('Unauthenticated. Authentication is required to remove a document from a collection'));
+        }
+
+        if ($user->cannot('update', $document) || $user->cannot('view', $collection)) {
+            throw new AuthorizationException(__('User not allowed to remove document from collection'));
+        }
 
 
         $collection->documents()->detach($document->getKey());
